changed gasprice param to int

+++ generate_signature_and_submit.py
@@ -71,7 +71,7 @@ def submit_key(web3_eth, account, keysmanager_contract, signature, publkey, nonc
     web3_eth.eth.call(tx)
     tx['nonce'] = nonce
     tx['gas'] = web3_eth.eth.estimate_gas(tx)
-    tx['gasPrice'] = 1.5*web3_eth.eth.generate_gas_price(tx)
+    tx['gasPrice'] = int(1.5*web3_eth.eth.generate_gas_price(tx))
     exit()
     signed_tx = web3_eth.eth.account.sign_transaction(tx, account.key)
     tx_hash = web3_eth.eth.send_raw_transaction(signed_tx.rawTransaction)
